Distinguishing template from set of templates
