+++ circusort/io/__init__.py
@@ -7,7 +7,8 @@ from .configure import Configuration
 from .parse import find_parameters_path, parse_parameters_file, parse_parameters
 from .probe import generate_probe, save_probe, load_probe, get_probe
 from .trains import generate_trains, save_trains, list_trains, load_train, load_trains, get_train
-from .template import generate_templates, save_templates, list_templates, load_template, load_templates, get_template
+from .template import load_template, get_template
+from .templates import generate_templates, save_templates, list_templates, load_templates
 from .position import generate_position, save_position, get_position
 from .parameter import *
 from .cell import generate_cell, load_cell, get_cell
